Se cambia double a float para reducir consumo de memoria en criba

+++ B_QSieve/sieve.h
@@ -7,5 +7,5 @@
 * @return 1 si se encontro solucion o 0 si no se encuentra una solucion
 */
 int shanksTonelli(mpz_t n, mpz_t p, mpz_t r1, mpz_t r2);
-double *sievingNaive(qs_struct * qs_data, enum TypeSieving typeSieving);
+float *sievingNaive(qs_struct * qs_data, enum TypeSieving typeSieving);
 long * sieving(qs_struct * qs_data, unsigned long *length);
